Move `BundleExtensions.kt` to :app:ui:base
